0.0.5 Two big changes: Main project buildable by default, generator argument passed via command line and defaults available for any platform.
This means the generator needs to be passed to the cinstall script locally for each platform. (e.g. if I want to build a project, I must pass "-CG Xcode" on OSX or npm run cinstall -- -CG "Visual Studio 16 2019" instead of an easier npm run cinstall. That's fine for now if that's the sacrifice for cross platform installs.